Add Kenya Flag SVG and Refactor GridMap Component
- Introduced a new SVG file for the Kenya flag in the public directory.
- Deleted the outdated GridMap component and replaced it with a new implementation that includes improved type safety and modular structure.
- Added supporting files for color management and utility functions related to the GridMap.
- Enhanced the InfoPanel and other related components to utilize the new GridMap structure, ensuring a cohesive user experience.
- Updated country configuration for Kenya to reflect new map view settings and improved localization support.
