+++ packages/nosql/src/constants/gaul2-districts.ts
@@ -3,35 +3,35 @@
 // ---------------------------------------------------------------------------
 
 
-export const GAUL2_DISTRICT_NAMES = [
-  "Chake Chake",
-  "Kaskazini A",
-  "Kaskazini B",
-  "Kati",
-  "Kusini",
-  "Magharibi A",
-  "Magharibi B",
-  "Micheweni",
-  "Mjini",
-  "Mkoani",
-  "Wete",
-] as const;
+// export const GAUL2_DISTRICT_NAMES = [
+//   "Chake Chake",
+//   "Kaskazini A",
+//   "Kaskazini B",
+//   "Kati",
+//   "Kusini",
+//   "Magharibi A",
+//   "Magharibi B",
+//   "Micheweni",
+//   "Mjini",
+//   "Mkoani",
+//   "Wete",
+// ] as const;
 
-export type GAUL2DistrictName = (typeof GAUL2_DISTRICT_NAMES)[number];
+// export type GAUL2DistrictName = (typeof GAUL2_DISTRICT_NAMES)[number];
 
-export const GAUL2_TO_REGION: Record<string, string> = {
-  "Chake Chake": "Pemba",
-  "Kaskazini A": "Unguja",
-  "Kaskazini B": "Unguja",
-  "Kati": "Unguja",
-  "Kusini": "Unguja",
-  "Magharibi A": "Unguja",
-  "Magharibi B": "Unguja",
-  "Micheweni": "Pemba",
-  "Mjini": "Unguja",
-  "Mkoani": "Pemba",
-  "Wete": "Pemba",
-};
+// export const GAUL2_TO_REGION: Record<string, string> = {
+//   "Chake Chake": "Pemba",
+//   "Kaskazini A": "Unguja",
+//   "Kaskazini B": "Unguja",
+//   "Kati": "Unguja",
+//   "Kusini": "Unguja",
+//   "Magharibi A": "Unguja",
+//   "Magharibi B": "Unguja",
+//   "Micheweni": "Pemba",
+//   "Mjini": "Unguja",
+//   "Mkoani": "Pemba",
+//   "Wete": "Pemba",
+// };
 
 
 // ---------------------------------------------------------------------------
@@ -39,49 +39,49 @@ export const GAUL2_TO_REGION: Record<string, string> = {
 // ---------------------------------------------------------------------------
 
 
-// export const GAUL2_DISTRICT_NAMES = [
-//   "Changamwe",
-//   "Jomvu",
-//   "Kilifi North",
-//   "Kilifi South",
-//   "Kinango",
-//   "Kisauni",
-//   "Lamu",
-//   "Lamu East",
-//   "Lamu West",
-//   "Likoni",
-//   "Lunga Lunga",
-//   "Magarini",
-//   "Malindi",
-//   "Matuga",
-//   "Msambweni",
-//   "Mvita",
-//   "Nyali",
-//   "Garsen",
-// ] as const;
+export const GAUL2_DISTRICT_NAMES = [
+  "Changamwe",
+  "Jomvu",
+  "Kilifi North",
+  "Kilifi South",
+  "Kinango",
+  "Kisauni",
+  "Lamu",
+  "Lamu East",
+  "Lamu West",
+  "Likoni",
+  "Lunga Lunga",
+  "Magarini",
+  "Malindi",
+  "Matuga",
+  "Msambweni",
+  "Mvita",
+  "Nyali",
+  "Garsen",
+] as const;
 
-// export type GAUL2DistrictName = (typeof GAUL2_DISTRICT_NAMES)[number];
+export type GAUL2DistrictName = (typeof GAUL2_DISTRICT_NAMES)[number];
 
-// export const GAUL2_TO_REGION: Record<string, string> = {
-//   'Changamwe': 'Central',
-//   'Jomvu': 'Central',
-//   'Kisauni': 'Central',
-//   'Likoni': 'Central',
-//   'Mvita': 'Central',
-//   'Nyali': 'Central',
-//   'Kilifi North': 'North',
-//   'Kilifi South': 'North',
-//   'Magarini': 'North',
-//   'Malindi': 'North',
-//   'Garsen': 'North',
-//   'Lamu': 'North',
-//   'Lamu East': 'North',
-//   'Lamu West': 'North',
-//   'Kinango': 'South',
-//   'Lunga Lunga': 'South',
-//   'Matuga': 'South',
-//   'Msambweni': 'South',
-// };
+export const GAUL2_TO_REGION: Record<string, string> = {
+  'Changamwe': 'Central',
+  'Jomvu': 'Central',
+  'Kisauni': 'Central',
+  'Likoni': 'Central',
+  'Mvita': 'Central',
+  'Nyali': 'Central',
+  'Kilifi North': 'North',
+  'Kilifi South': 'North',
+  'Magarini': 'North',
+  'Malindi': 'North',
+  'Garsen': 'North',
+  'Lamu': 'North',
+  'Lamu East': 'North',
+  'Lamu West': 'North',
+  'Kinango': 'South',
+  'Lunga Lunga': 'South',
+  'Matuga': 'South',
+  'Msambweni': 'South',
+};
 
 
 // ---------------------------------------------------------------------------
